data_flow_diagram_microservice/requirements.txt
pyautogen==0.2.0b3
fastapi==0.104.1
uvicorn==0.24.0.post1
pytm==1.2.1
graphviz==0.20.1
pdf-reports==0.3.5
weasyprint==60.1
matplotlib==3.8.1
